requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.3
pydantic-settings==2.1.0

# Environment & Config
python-dotenv==1.0.0

# Twilio for SMS
twilio==8.12.0

# Utilities
requests==2.31.0

# Optional: For better performance
gunicorn==21.2.0